# Fix up the Layout Manager