This PR:

- Adds some `LayoutManager` unit tests that were failing on `master`
- Makes those pass
- Adds a layout benchmark
- Tweaks the layout algorithm for the best benchmark result

Various layout algorithms were experimented with in the tweaks of the layout algorithm, they are presented below with benchmarks:

## [Tidied up algorithm from `master`](ac3ca7ca292d4e59ca4f373ee04918a8d6da0ee5)

5ms 156.25 Gen0 481.64kb total

## [DirtyVisuals](c48bc34ef9b78c229ee1119f6b0f3f87dfd26f0d)

The `DirtyVisuals` class was ported from `scenegraph` and used to track invalidated controls.

5.1ms 195.31 Gen 0 604kb total

## [DistanceFromRoot](927f085aee73b7059cc60c1fc12510b32860c426)

Following on from DirtyVisuals, a `Visual.DistanceFromRoot` property was added and `DirtyVisuals` modified to use that.

4.78ms 195 Gen0 604kb total

## [List binary search](c500d22c84e95db9b6091378a1f87256d825de02)

`DirtyVisuals` was modified to use a simple `List` ordered by `DistanceFromRoot` and use a binary search to insert items.

4.611ms 148 Gen0 463kb total

## Current implementation

The previous changes were undone to ac3ca7ca292d4e59ca4f373ee04918a8d6da0ee5 and `LayoutManager` was then modified to use a `Queue` instead of a `HasSet` as @kekekeks pointed out that a duplicate check should be unnecessary as a control should not invalidate itself if it's already invalidated.

4.028ms 148.437 Gen0 463kb total