@grokys 2017-06-12
@kekekeks implemented the changes we spoke about in gitter. We don't actually need to explicitly invalidate a control with the layout manager when its added to the tree - [that's already taken care of](https://github.com/AvaloniaUI/Avalonia/pull/1021/commits/40c342989b538022a82d6c12151ea1995499d44f#diff-66d78698a4e8a841097dbf0985cc7e77R53)